requirements.txt
Pillow==2.6.1
click==4.0
coverage==4.0a5
images2gif==1.0.1
mock==1.0.1
nose==1.3.6
numpy==1.9.2
oauthlib==0.7.2
python-dateutil==2.4.2
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
twython==3.2.0
wsgiref==0.1.2
